edit: priority working

+++ app/src/main/java/hu/ait/tovisitmapapp/data/ToVisitItem.kt
@@ -12,7 +12,7 @@ data class ToVisitItem(
     @PrimaryKey(autoGenerate = true) var id: Int = 0,
     @ColumnInfo(name = "name") var name: String,
     @ColumnInfo(name = "description") var description: String,
-    @ColumnInfo(name = "priority") var priority: Int,
+    @ColumnInfo(name = "priority") var priority: Float,
     @ColumnInfo(name = "category") var category: ToVisitCategory,
     @ColumnInfo(name = "havevisited") var haveVisited: Boolean,
     @ColumnInfo(name = "address") var address: String,
